+++ apps/frontend/src/libs/types/quiz-components-properties.type.ts
@@ -1,8 +1,5 @@
 import { type QuestionDto } from "~/libs/types/types.js";
-import {
-	type QuizAnswer,
-	type QuizCategoryValue,
-} from "~/modules/quiz/quiz.js";
+import { type QuizAnswer } from "~/modules/quiz/quiz.js";
 
 type CheckboxQuestionProperties = {
 	currentAnswer?: MultipleAnswers[];
@@ -50,10 +47,10 @@ type QuizCategoryCardProperties = QuizCategoryProperties & {
 };
 
 type QuizCategoryProperties = {
-	category: QuizCategoryValue;
 	color: string;
-	icon: string;
+	iconHref: string;
 	selected: boolean;
+	title: string;
 };
 
 type RadioQuestionProperties = {
@@ -90,7 +87,6 @@ export {
 	type QuestionNavigationProperties,
 	type QuestionPageProperties,
 	type QuizCategoryCardProperties,
-	type QuizCategoryProperties,
 	type RadioQuestionProperties,
 	type SingleAnswer,
 	type SingleChoiceWithTextAnswer,