+++ apps/frontend/src/modules/plans/slices/actions.ts
@@ -1,22 +1,27 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 
+import { StorageKey } from "~/libs/modules/storage/storage.js";
 import { type AsyncThunkConfig } from "~/libs/types/types.js";
 import { type PlanDaysTaskDto } from "~/modules/plans/plans.js";
 
 import {
+	type GeneratePlanRequestDto,
 	type PlanSearchQueryParameter,
 	type PlanWithCategoryDto,
-	type QuizAnswersRequestDto,
 } from "../libs/types/types.js";
 import { name as sliceName } from "./plan.slice.js";
 
 const generatePlan = createAsyncThunk<
 	PlanDaysTaskDto,
-	QuizAnswersRequestDto,
+	GeneratePlanRequestDto,
 	AsyncThunkConfig
 >(`${sliceName}/generate`, async (payload, { extra }) => {
-	const { planApi } = extra;
+	const { planApi, storage } = extra;
 	const plan = await planApi.generate(payload);
+
+	if (!plan.userId) {
+		await storage.set(StorageKey.PLAN_ID, String(plan.id));
+	}
 
 	return plan;
 });
@@ -43,4 +48,14 @@ const getAllUserPlans = createAsyncThunk<
 	return plan;
 });
 
-export { generatePlan, getAllUserPlans, searchPlan };
+const findPlan = createAsyncThunk<PlanDaysTaskDto, number, AsyncThunkConfig>(
+	`${sliceName}/:id`,
+	async (payload, { extra }) => {
+		const { planApi } = extra;
+		const plan = await planApi.findWithRelations(payload);
+
+		return plan;
+	},
+);
+
+export { findPlan, generatePlan, getAllUserPlans, searchPlan };