+++ apps/frontend/src/modules/plans/slices/plan.slice.ts
@@ -1,14 +1,20 @@
-import { createSlice } from "@reduxjs/toolkit";
+import { createSlice, type PayloadAction } from "@reduxjs/toolkit";
 
-import { DataStatus } from "~/libs/enums/enums.js";
-import { type ValueOf } from "~/libs/types/types.js";
+import { DataStatus, PlanStyle } from "~/libs/enums/enums.js";
+import { type PlanStyleOption, type ValueOf } from "~/libs/types/types.js";
 import { type PlanDaysTaskDto } from "~/modules/plans/plans.js";
 
-import { generatePlan, getAllUserPlans, searchPlan } from "./actions.js";
+import {
+	findPlan,
+	generatePlan,
+	getAllUserPlans,
+	searchPlan,
+} from "./actions.js";
 
 type State = {
 	dataStatus: ValueOf<typeof DataStatus>;
 	plan: null | PlanDaysTaskDto;
+	selectedStyle: PlanStyleOption;
 	userPlans: PlanDaysTaskDto[];
 	userPlansDataStatus: ValueOf<typeof DataStatus>;
 };
@@ -16,6 +22,7 @@ type State = {
 const initialState: State = {
 	dataStatus: DataStatus.IDLE,
 	plan: null,
+	selectedStyle: PlanStyle.WITH_REMARKS,
 	userPlans: [],
 	userPlansDataStatus: DataStatus.IDLE,
 };
@@ -55,10 +62,28 @@ const { actions, name, reducer } = createSlice({
 			state.userPlansDataStatus = DataStatus.REJECTED;
 			state.userPlans = [];
 		});
+		builder.addCase(findPlan.pending, (state) => {
+			state.userPlansDataStatus = DataStatus.PENDING;
+		});
+		builder.addCase(findPlan.fulfilled, (state, action) => {
+			state.userPlansDataStatus = DataStatus.FULFILLED;
+			state.plan = action.payload;
+		});
+		builder.addCase(findPlan.rejected, (state) => {
+			state.userPlansDataStatus = DataStatus.REJECTED;
+			state.plan = null;
+		});
 	},
 	initialState,
 	name: "plan",
-	reducers: {},
+	reducers: {
+		setPlan: (state, action: PayloadAction<PlanDaysTaskDto>) => {
+			state.plan = action.payload;
+		},
+		setSelectedStyle: (state, action: PayloadAction<PlanStyleOption>) => {
+			state.selectedStyle = action.payload;
+		},
+	},
 });
 
 export { actions, name, reducer };