+++ apps/frontend/src/pages/dashboard-wrapper-mock/components/plan/components/task/task.tsx
@@ -23,6 +23,7 @@ type Properties = {
 	indexItem: number;
 	item: {
 		description: string;
+		executionTimeType: string;
 		id: number;
 		title: string;
 	};
@@ -31,51 +32,40 @@ type Properties = {
 const Task: React.FC<Properties> = ({ indexItem, item }: Properties) => {
 	const [isEditing, setIsEditing] = useState<boolean>(false);
 	const [editedTitle, setEditedTitle] = useState<string>(item.title);
-	const [editedDescription, setEditedDescription] = useState<string>(
-		item.description,
-	);
+
 	const [isDeleteModalOpen, setIsDeleteModalOpen] = useState<boolean>(false);
 
 	const dispatch = useAppDispatch();
 
 	useEffect(() => {
 		setEditedTitle(item.title);
-		setEditedDescription(item.description);
-	}, [item.title, item.description]);
+	}, [item.title]);
 
 	const handleEditClick = useCallback((): void => {
 		if (isEditing) {
 			setEditedTitle(item.title);
-			setEditedDescription(item.description);
 			setIsEditing(false);
 		} else {
 			setIsEditing(true);
 		}
-	}, [isEditing, item.title, item.description]);
+	}, [isEditing, item.title]);
 
 	const handleSaveClick = useCallback((): void => {
 		void dispatch(
 			taskActions.updateTask({
 				id: item.id,
 				payload: {
-					description: editedDescription,
+					description: item.description,
 					title: editedTitle,
 				},
 			}),
 		);
 		setIsEditing(false);
-	}, [dispatch, item.id, editedDescription, editedTitle]);
+	}, [dispatch, item.id, item.description, editedTitle]);
 
 	const handleTitleChange = useCallback(
 		(event: React.ChangeEvent<HTMLInputElement>): void => {
 			setEditedTitle(event.target.value);
-		},
-		[],
-	);
-
-	const handleDescriptionChange = useCallback(
-		(event: React.ChangeEvent<HTMLTextAreaElement>): void => {
-			setEditedDescription(event.target.value);
 		},
 		[],
 	);
@@ -122,20 +112,6 @@ const Task: React.FC<Properties> = ({ indexItem, item }: Properties) => {
 								{editedTitle.length}/{TaskValidationRule.TITLE_MAX_LENGTH}{" "}
 								characters
 							</div>
-							<textarea
-								className={getClassNames(
-									styles["edit-input"],
-									styles["edit-input-description"],
-								)}
-								maxLength={TaskValidationRule.DESCRIPTION_MAX_LENGTH}
-								onChange={handleDescriptionChange}
-								placeholder="Task description"
-								value={editedDescription}
-							/>
-							<div className={styles["char-counter"]}>
-								{editedDescription.length}/
-								{TaskValidationRule.DESCRIPTION_MAX_LENGTH} characters
-							</div>
 							<div className={styles["save-button-wrapper"]}>
 								<Button
 									className={getClassNames(styles["save-button"])}
@@ -157,7 +133,7 @@ const Task: React.FC<Properties> = ({ indexItem, item }: Properties) => {
 				<div className={styles["item-actions"]}>
 					<div className={styles["item-actions__time"]}>
 						<img alt="Timer" src={Timer} />
-						<span>morning</span>
+						<span>{item.executionTimeType}</span>
 					</div>
 					<div className={styles["item-actions_buttons-wrapper"]}>
 						<Button