+++ apps/frontend/src/routes.tsx
@@ -16,6 +16,7 @@ import {
 	QuestionFlow,
 	Quiz,
 } from "./pages/pages.js";
+import { PlanEdit } from "./pages/plan-edit/plan-edit.js";
 import { TestPage } from "./pages/test-page/test-page.js";
 
 type CustomRouteObject = RouteObject & { handle: RouteHandle };
@@ -56,6 +57,11 @@ const routes: CustomRouteObject[] = [
 		path: AppRoute.QUIZ_QUESTIONS,
 	},
 	{
+		element: <PlanEdit />,
+		handle: { access: RouteAccess.PUBLIC },
+		path: AppRoute.PLAN_EDIT,
+	},
+	{
 		element: <Quiz />,
 		handle: { access: RouteAccess.PUBLIC },
 		path: AppRoute.QUIZ,
@@ -74,7 +80,7 @@ const routes: CustomRouteObject[] = [
 			},
 			{
 				element: <Plan />,
-				handle: { access: RouteAccess.PUBLIC },
+				handle: { access: RouteAccess.AUTHENTICATED },
 				path: AppRoute.PLAN,
 			},
 		],